Moved some shared stuff to utils.py
